+++ src/ResXManager.Scripting/Host.cs
@@ -198,7 +198,11 @@
 
         public string? EffectiveTranslationPrefix { get; set; }
 
-        public PrefixFieldType PrefixFieldType { get; set; }
+        public bool PrefixValue { get; set; }
+
+        public bool PrefixNeutralComment { get; set; }
+
+        public bool PrefixTargetComment { get; set; }
 
         public ExcelExportMode ExcelExportMode { get; set; }
 
@@ -211,6 +215,8 @@
         public string? TranslatorConfiguration { get; set; }
 
         public bool AutoApplyExistingTranslations { get; set; }
+
+        public string? CultureCountyOverrides { get; set; }
 
         public event PropertyChangedEventHandler? PropertyChanged;